backend/requirements.txt
Django==4.2.11
djangorestframework
djangorestframework-simplejwt
django-cors-headers

channels

gunicorn
whitenoise

psycopg2-binary
Pillow

qrcode
requests
python-dotenv

channels[daphne]
daphne
uvicorn[standard]

cvzone
imageio-ffmpeg
numpy>=1.26.4,<2.0
ultralytics
opencv-python
lap
easyocr



cloudinary
django-cloudinary-storage
dj-database-url
